Function to assign population to centers of services

Function to process the results from EPP::clust_it

assign_clust(clustered)

Arguments

clustered
List of the population assigned to each center by EPP::clust_it

Value

list with Centers of the clusters with the number of individuals covered for each service distance and mean distances and Population with the center assigned to

Examples

a <- clust_it(pop_epp)
a <- assign_clust(clustered = a)

#> Error in (function (..., row.names = NULL, check.rows = FALSE, check.names = TRUE,     fix.empty.names = TRUE, stringsAsFactors = default.stringsAsFactors()) {    data.row.names <- if (check.rows && is.null(row.names))         function(current, new, i) {            if (is.character(current))                 new <- as.character(new)            if (is.character(new))                 current <- as.character(current)            if (anyDuplicated(new))                 return(current)            if (is.null(current))                 return(new)            if (all(current == new) || all(current == ""))                 return(new)            stop(gettextf("mismatch of row names in arguments of 'data.frame', item %d",                 i), domain = NA)        }    else function(current, new, i) {        if (is.null(current)) {            if (anyDuplicated(new)) {                warning(gettextf("some row.names duplicated: %s --> row.names NOT used",                   paste(which(duplicated(new)), collapse = ",")),                   domain = NA)                current            }            else new        }        else current    }    object <- as.list(substitute(list(...)))[-1L]    mirn <- missing(row.names)    mrn <- is.null(row.names)    x <- list(...)    n <- length(x)    if (n < 1L) {        if (!mrn) {            if (is.object(row.names) || !is.integer(row.names))                 row.names <- as.character(row.names)            if (anyNA(row.names))                 stop("row names contain missing values")            if (anyDuplicated(row.names))                 stop(gettextf("duplicate row.names: %s", paste(unique(row.names[duplicated(row.names)]),                   collapse = ", ")), domain = NA)        }        else row.names <- integer()        return(structure(list(), names = character(), row.names = row.names,             class = "data.frame"))    }    vnames <- names(x)    if (length(vnames) != n)         vnames <- character(n)    no.vn <- !nzchar(vnames)    vlist <- vnames <- as.list(vnames)    nrows <- ncols <- integer(n)    for (i in seq_len(n)) {        xi <- if (is.character(x[[i]]) || is.list(x[[i]]))             as.data.frame(x[[i]], optional = TRUE, stringsAsFactors = stringsAsFactors)        else as.data.frame(x[[i]], optional = TRUE)        nrows[i] <- .row_names_info(xi)        ncols[i] <- length(xi)        namesi <- names(xi)        if (ncols[i] > 1L) {            if (length(namesi) == 0L)                 namesi <- seq_len(ncols[i])            vnames[[i]] <- if (no.vn[i])                 namesi            else paste(vnames[[i]], namesi, sep = ".")        }        else if (length(namesi)) {            vnames[[i]] <- namesi        }        else if (fix.empty.names && no.vn[[i]]) {            tmpname <- deparse(object[[i]], nlines = 1L)[1L]            if (substr(tmpname, 1L, 2L) == "I(") {                ntmpn <- nchar(tmpname, "c")                if (substr(tmpname, ntmpn, ntmpn) == ")")                   tmpname <- substr(tmpname, 3L, ntmpn - 1L)            }            vnames[[i]] <- tmpname        }        if (mirn && nrows[i] > 0L) {            rowsi <- attr(xi, "row.names")            if (any(nzchar(rowsi)))                 row.names <- data.row.names(row.names, rowsi,                   i)        }        nrows[i] <- abs(nrows[i])        vlist[[i]] <- xi    }    nr <- max(nrows)    for (i in seq_len(n)[nrows < nr]) {        xi <- vlist[[i]]        if (nrows[i] > 0L && (nr%%nrows[i] == 0L)) {            xi <- unclass(xi)            fixed <- TRUE            for (j in seq_along(xi)) {                xi1 <- xi[[j]]                if (is.vector(xi1) || is.factor(xi1))                   xi[[j]] <- rep(xi1, length.out = nr)                else if (is.character(xi1) && inherits(xi1, "AsIs"))                   xi[[j]] <- structure(rep(xi1, length.out = nr),                     class = class(xi1))                else if (inherits(xi1, "Date") || inherits(xi1,                   "POSIXct"))                   xi[[j]] <- rep(xi1, length.out = nr)                else {                  fixed <- FALSE                  break                }            }            if (fixed) {                vlist[[i]] <- xi                next            }        }        stop(gettextf("arguments imply differing number of rows: %s",             paste(unique(nrows), collapse = ", ")), domain = NA)    }    value <- unlist(vlist, recursive = FALSE, use.names = FALSE)    vnames <- unlist(vnames[ncols > 0L])    if (fix.empty.names && any(noname <- !nzchar(vnames)))         vnames[noname] <- paste0("Var.", seq_along(vnames))[noname]    if (check.names) {        if (fix.empty.names)             vnames <- make.names(vnames, unique = TRUE)        else {            nz <- nzchar(vnames)            vnames[nz] <- make.names(vnames[nz], unique = TRUE)        }    }    names(value) <- vnames    if (!mrn) {        if (length(row.names) == 1L && nr != 1L) {            if (is.character(row.names))                 row.names <- match(row.names, vnames, 0L)            if (length(row.names) != 1L || row.names < 1L ||                 row.names > length(vnames))                 stop("'row.names' should specify one of the variables")            i <- row.names            row.names <- value[[i]]            value <- value[-i]        }        else if (!is.null(row.names) && length(row.names) !=             nr)             stop("row names supplied are of the wrong length")    }    else if (!is.null(row.names) && length(row.names) != nr) {        warning("row names were found from a short variable and have been discarded")        row.names <- NULL    }    class(value) <- "data.frame"    if (is.null(row.names))         attr(value, "row.names") <- .set_row_names(nr)    else {        if (is.object(row.names) || !is.integer(row.names))             row.names <- as.character(row.names)        if (anyNA(row.names))             stop("row names contain missing values")        if (anyDuplicated(row.names))             stop(gettextf("duplicate row.names: %s", paste(unique(row.names[duplicated(row.names)]),                 collapse = ", ")), domain = NA)        row.names(value) <- row.names    }    value})(structure(list(cluster = structure(c(1L, 1L, 1L, 10L, 10L, 10L, 10L, 11L, 11L, 11L, 11L, 12L, 12L, 12L, 12L, 12L, 13L, 13L, 13L, 14L, 2L, 3L, 3L, 3L, 4L, 4L, 4L, 4L, 4L, 4L, 5L, 5L, 6L, 6L, 6L, 6L, 6L, 7L, 8L, 8L, 8L, 8L, 8L, 9L, 9L, 9L), .Label = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14"), class = "factor"), x = c(577415L, 578467L, 576666L, 573373L, 573422L, 573020L, 574392L, 583719L, 583856L, 583835L, 584404L, 570092L, 569899L, 570600L, 570140L, 570584L, 577526L, 577620L, 576877L, 580669L, 562835L, 576179L, 575861L, 576661L, 566368L, 566549L, 566527L, 566793L, 566404L, 566102L, 561361L, 561234L, 569741L, 569636L, 569571L, 570747L, 569530L, 583705L, 568784L, 569875L, 569943L, 569058L, 568603L, 563676L, 564266L, 563400L), y = c(6156711L, 6156522L, 6156619L, 6151711L, 6150979L, 6151076L, 6151932L, 6143298L, 6143095L, 6142400L, 6142446L, 6148389L, 6147976L, 6148401L, 6148806L, 6148968L, 6137867L, 6138233L, 6138269L, 6149557L, 6148314L, 6144184L, 6143623L, 6144029L, 6153287L, 6152922L, 6152586L, 6153638L, 6153792L, 6153902L, 6156823L, 6156035L, 6140546L, 6141010L, 6140123L, 6140693L, 6139988L, 6157576L, 6157593L, 6158022L, 6157893L, 6158659L, 6158395L, 6139592L, 6138945L, 6139893L), weight = c(0.911494901403785, 0.973810895811766, 0.76299475133419, 0.258552481653169, 0.119053601752967, 0.249491078080609, 0.841804659226909, 0.758026976836845, 0.924203391885385, 0.349126908229664, 0.500924616120756, 0.680366796907037, 0.535781325306743, 0.107969269156456, 0.873356754891574, 0.33490219945088, 0.180040644248948, 0.390199792804196, 0.169575409730896, 0.0545469161588699, 0.532091336790472, 0.500427180668339, 0.638665714534, 0.0848551793023944, 0.215378106571734, 0.808228050358593, 0.817400673404336, 0.982604091288522, 0.545283400686458, 0.881442637415603, 0.155597463948652, 0.528351496905088, 0.243636747589335, 0.683684614254162, 0.43789167678915, 0.168518107384443, 0.604005304630846, 0.502442672848701, 0.207585948752239, 0.433079183334485, 0.805740788811818, 0.375521864742041, 0.0489185054320842, 0.279622281203046, 0.292433400405571, 0.123046519933268), medianx = c(577600.5, 577600.5, 577600.5, 573612, 573612, 573612, 573612, 583777, 583777, 583777, 583777, 570140, 570140, 570140, 570140, 570140, 577653.5, 577653.5, 577653.5, 581089.5, 562739, 575880, 575880, 575880, 566435, 566435, 566435, 566435, 566435, 566435, 561929, 561929, 570016, 570016, 570016, 570016, 570016, 582380, 569350.5, 569350.5, 569350.5, 569350.5, 569350.5, 563676, 563676, 563676), mediany = c(6156901.5, 6156901.5, 6156901.5, 6151627.5, 6151627.5, 6151627.5, 6151627.5, 6143196.5, 6143196.5, 6143196.5, 6143196.5, 6148200, 6148200, 6148200, 6148200, 6148200, 6137681, 6137681, 6137681, 6148682, 6147527.5, 6144029, 6144029, 6144029, 6153152.5, 6153152.5, 6153152.5, 6153152.5, 6153152.5, 6153152.5, 6156429, 6156429, 6140678, 6140678, 6140678, 6140678, 6140678, 6157565, 6157743, 6157743, 6157743, 6157743, 6157743, 6139166, 6139166, 6139166), dist = c(265.895656226272, 945.961151422192, 976.266613174905, 253.166447223956, 675.760497513727, 809.083586534791, 837.329236322249, 116.902737350329, 128.620566007151, 798.608946856971, 977.946445363958, 195, 329.024315210897, 501.997011943298, 606, 887.107659757259, 225.504434546197, 553.015596525089, 974.010395221735, 970.796193853272, 792.337207254588, 336.787766998744, 406.444338132542, 781, 150.263934461999, 257.150247909661, 573.921815232702, 603.21990186001, 640.250927371449, 820.145871171708, 691.274185833668, 798.912385684438, 305.039341724965, 504.602814102339, 711.371913980303, 731.153882571925, 843.976302984865, 1325.04565959064, 586.022397182906, 594.088587670223, 611.19248195638, 961.567600327715, 991.897298111049, 426, 630.032538842241, 777.627803000896), orden_dist = c(1L, 2L, 3L, 1L, 2L, 3L, 4L, 1L, 2L, 3L, 4L, 1L, 2L, 3L, 4L, 5L, 1L, 2L, 3L, 1L, 1L, 1L, 2L, 3L, 1L, 2L, 3L, 4L, 5L, 6L, 1L, 2L, 1L, 2L, 3L, 4L, 5L, 1L, 1L, 2L, 3L, 4L, 5L, 1L, 2L, 3L), max_n_cl = c(3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 3, 3, 3, 1, 1, 3, 3, 3, 6, 6, 6, 6, 6, 6, 2, 2, 5, 5, 5, 5, 5, 1, 5, 5, 5, 5, 5, 3, 3, 3), a_reasig = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), round = structure(c(1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L), class = "factor", .Label = "1")), row.names = c(1L, 2L, 3L, 37L, 38L, 39L, 40L, 73L, 74L, 75L, 76L, 97L, 98L, 99L, 100L, 101L, 144L, 145L, 146L, 178L, 214L, 250L, 251L, 252L, 291L, 292L, 293L, 294L, 295L, 296L, 321L, 322L, 355L, 356L, 357L, 358L, 359L, 403L, 424L, 425L, 426L, 427L, 428L, 458L, 459L, 460L), class = "data.frame"),     structure(list(cluster = structure(c(1L, 10L, 10L, 10L, 10L,     11L, 12L, 12L, 13L, 2L, 2L, 2L, 3L, 3L, 3L, 3L, 4L, 4L, 4L,     4L, 5L, 5L, 5L, 5L, 5L, 6L, 6L, 7L, 7L, 7L, 8L, 9L), .Label = c("1",     "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12",     "13"), class = "factor"), x = c(581048L, 573490L, 573112L,     573089L, 573598L, 568254L, 561681L, 562577L, 582356L, 574582L,     574777L, 573674L, 577214L, 577000L, 577515L, 577696L, 562917L,     562994L, 562493L, 563301L, 565122L, 564341L, 565933L, 565050L,     565548L, 581546L, 580940L, 572535L, 572679L, 572641L, 570373L,     564768L), y = c(6157335L, 6139514L, 6139411L, 6139480L, 6138169L,     6155603L, 6157622L, 6157817L, 6138526L, 6155535L, 6155278L,     6155222L, 6143868L, 6144225L, 6143477L, 6144070L, 6143418L,     6142969L, 6143182L, 6143326L, 6137589L, 6137782L, 6137975L,     6138648L, 6138543L, 6147414L, 6147519L, 6150004L, 6150196L,     6150281L, 6144114L, 6149954L), weight = c(0.702034631045535,     0.845292892307043, 0.0842704367823899, 0.43241948960349,     0.384575875243172, 0.539254390634596, 0.098392520332709,     0.541174229700118, 0.177937475033104, 0.960196747910231,     0.345192675245926, 0.281362429261208, 0.0384490031283349,     0.530519535532221, 0.682035467354581, 0.819946319796145,     0.959073857171461, 0.0246966008562595, 0.930702910991386,     0.690849122591317, 0.754853080725297, 0.150647377362475,     0.233002191642299, 0.689002977916971, 0.89547887374647, 0.816140659619123,     0.0440494054928422, 0.671137500088662, 0.0902555836364627,     0.29360425285995, 0.928075806237757, 0.0723129622638226),         medianx = c(580692, 573598, 573598, 573598, 573598, 568684,         562113, 562113, 581582, 574625, 574625, 574625, 576768,         576768, 576768, 576768, 562798, 562798, 562798, 562798,         565122, 565122, 565122, 565122, 565122, 581625.5, 581625.5,         572184, 572184, 572184, 569357.5, 564761), mediany = c(6157029,         6139046, 6139046, 6139046, 6139046, 6156524, 6157201,         6157201, 6138605.5, 6155496, 6155496, 6155496, 6143814.5,         6143814.5, 6143814.5, 6143814.5, 6143182, 6143182, 6143182,         6143182, 6137782, 6137782, 6137782, 6137782, 6137782,         6147967.5, 6147967.5, 6149501, 6149501, 6149501, 6144271,         6149922), dist = c(469.437961822433, 480.299906308548,         607.80013162223, 668.907317944721, 877, 1016.43543818582,         603.212234623934, 771.201659749251, 778.07213675854,         58.0517010948, 265.759289583638, 989.685303518245, 449.197339707171,         471.523329221365, 819.704367439872, 962.530129398555,         264.304748349325, 289.456387043022, 305, 523.206460204765,         193, 781, 833.648607028165, 868.987917062142, 872.122124475695,         559.180203512249, 819.184045254789, 613.359600886788,         853.258460256914, 904.018252028132, 1027.56471815648,         32.7566787083184), orden_dist = c(1L, 1L, 2L, 3L, 4L,         1L, 1L, 2L, 1L, 1L, 2L, 3L, 1L, 2L, 3L, 4L, 1L, 2L, 3L,         4L, 1L, 2L, 3L, 4L, 5L, 1L, 2L, 1L, 2L, 3L, 1L, 1L),         max_n_cl = c(1, 4, 4, 4, 4, 1, 2, 2, 1, 3, 3, 3, 4, 4,         4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 2, 2, 3, 3, 3, 1, 1),         a_reasig = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),         round = structure(c(1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L), class = "factor", .Label = "2")), row.names = c(1L,     40L, 41L, 42L, 43L, 75L, 110L, 111L, 141L, 165L, 166L, 167L,     201L, 202L, 203L, 204L, 233L, 234L, 235L, 236L, 262L, 263L,     264L, 265L, 266L, 293L, 294L, 335L, 336L, 337L, 374L, 412L    ), class = "data.frame"), structure(list(cluster = structure(c(1L,     10L, 10L, 11L, 12L, 12L, 12L, 2L, 3L, 4L, 5L, 5L, 6L, 7L,     7L, 7L, 7L, 7L, 7L, 7L, 8L, 8L, 8L, 8L, 9L, 9L, 9L), .Label = c("1",     "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12"), class = "factor"),         x = c(575428L, 573304L, 573882L, 574697L, 561918L, 562173L,         561462L, 568186L, 581166L, 562413L, 580852L, 580185L,         582404L, 568890L, 568510L, 568253L, 568222L, 568274L,         569393L, 568299L, 567941L, 566953L, 568017L, 566505L,         561703L, 562643L, 561773L), y = c(6156099L, 6147456L,         6146057L, 6138716L, 6153678L, 6153462L, 6153467L, 6150681L,         6156116L, 6137531L, 6140105L, 6141641L, 6148675L, 6143197L,         6143091L, 6143318L, 6143411L, 6142875L, 6142992L, 6142476L,         6157372L, 6158026L, 6157381L, 6157296L, 6146377L, 6145821L,         6146041L), weight = c(0.966786335688084, 0.398337963037193,         0.807603048859164, 0.533486451255158, 0.0334391670767218,         0.765638109995052, 0.766856846632436, 0.467672932427377,         0.87010829965584, 0.779595609055832, 0.914116001222283,         0.464897213736549, 0.626434272620827, 0.71442348160781,         0.329071169253439, 0.228875158820301, 0.370325124822557,         0.630127151496708, 0.290813800180331, 0.0258299454580992,         0.615122395101935, 0.991382177686319, 0.759243806358427,         0.113095779204741, 0.661699079442769, 0.816063771722838,         0.809361269231886), medianx = c(574668, 573933.5, 573933.5,         574304.5, 562147, 562147, 562147, 568979, 580623, 563349.5,         580704.5, 580704.5, 581358, 568760, 568760, 568760, 568760,         568760, 568760, 568760, 567419, 567419, 567419, 567419,         562453, 562453, 562453), mediany = c(6155457, 6146990,         6146990, 6139033.5, 6154152.5, 6154152.5, 6154152.5,         6150681, 6156957.5, 6138758, 6140816.5, 6140816.5, 6148192,         6143255, 6143255, 6143255, 6143255, 6143255, 6143255,         6143255, 6157651, 6157651, 6157651, 6157651, 6146726,         6146726, 6146726), dist = c(994.868835575826, 783.215327990968,         934.420274822844, 504.839083273076, 526.869291190899,         690.989326979802, 969.089908109666, 793, 1001.48452309559,         1543.55474473697, 726.628171763248, 974.51552065629,         1152.13063495421, 142.351677194194, 298.991638679077,         510.899207280653, 560.16069123065, 616.924630728908,         685.461888072561, 905.186168696805, 591.882589708466,         598.147975002842, 656.128036285602, 980.520779993979,         827.224878736127, 924.729690233854, 965.207231634741),         orden_dist = c(1L, 1L, 2L, 1L, 1L, 2L, 3L, 1L, 1L, 1L,         1L, 2L, 1L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 1L, 2L, 3L, 4L,         1L, 2L, 3L), max_n_cl = c(1, 2, 2, 1, 3, 3, 3, 1, 1,         1, 2, 2, 1, 7, 7, 7, 7, 7, 7, 7, 4, 4, 4, 4, 3, 3, 3),         a_reasig = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), round = structure(c(1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L), class = "factor", .Label = "3")), row.names = c(1L,     36L, 37L, 78L, 116L, 117L, 118L, 152L, 189L, 227L, 257L,     258L, 293L, 330L, 331L, 332L, 333L, 334L, 335L, 336L, 365L,     366L, 367L, 368L, 392L, 393L, 394L), class = "data.frame"),     structure(list(cluster = structure(c(1L, 10L, 11L, 11L, 2L,     2L, 3L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 5L, 5L, 5L, 6L, 6L, 7L,     7L, 8L, 8L, 9L), .Label = c("1", "2", "3", "4", "5", "6",     "7", "8", "9", "10", "11"), class = "factor"), x = c(574172L,     574461L, 561182L, 562133L, 565676L, 566692L, 567254L, 567302L,     567375L, 566762L, 566045L, 564928L, 562478L, 562224L, 562321L,     562624L, 561940L, 575565L, 576298L, 570039L, 570723L, 581314L,     581662L, 582453L), y = c(6147965L, 6156917L, 6144220L, 6145100L,     6139166L, 6138984L, 6148210L, 6148282L, 6148595L, 6148523L,     6159018L, 6159901L, 6152342L, 6152577L, 6153064L, 6151852L,     6152040L, 6139008L, 6139501L, 6153345L, 6153757L, 6154522L,     6155894L, 6143840L), weight = c(0.605649225413799, 0.308140227105469,     0.0130800546612591, 0.55391551903449, 0.221341451397166,     0.388444478390738, 0.699274405837059, 0.269445938291028,     0.00214724359102547, 0.403696686029434, 0.951503718504682,     0.500823911977932, 0.516792953712866, 0.797296215780079,     0.175955580314621, 0.197111350949854, 0.547555505065247,     0.502524690469727, 0.224759430624545, 0.989779067924246,     0.832114716991782, 0.231572725344449, 0.0187503222841769,     0.513669215142727), medianx = c(573888, 575215, 561546, 561546,     566002.5, 566002.5, 567338.5, 567338.5, 567338.5, 567338.5,     565308, 565308, 562478, 562478, 562478, 562478, 562478, 575736,     575736, 570532.5, 570532.5, 581094.5, 581094.5, 582453),         mediany = c(6146877, 6157320.5, 6144302, 6144302, 6138783,         6138783, 6147927, 6147927, 6147927, 6147927, 6159077,         6159077, 6152577, 6152577, 6152577, 6152577, 6152577,         6139046, 6139046, 6153152.5, 6153152.5, 6155171.5, 6155171.5,         6144911), dist = c(1124.45542374965, 855.177320793764,         373.121963974248, 990.642720661692, 503.280488395884,         718.200006961849, 295.345983551495, 356.871475464207,         668.996449915842, 829.197352866011, 739.357829470954,         907.400683270627, 235, 254, 511.681541586171, 739.554595685809,         760.140118662342, 175.171344688565, 723.09681232875,         529.715489673466, 633.80635844081, 685.58770409044, 918.728741250648,         1071), orden_dist = c(1L, 1L, 1L, 2L, 1L, 2L, 1L, 2L,         3L, 4L, 1L, 2L, 1L, 2L, 3L, 4L, 5L, 1L, 2L, 1L, 2L, 1L,         2L, 1L), max_n_cl = c(1, 1, 2, 2, 2, 2, 4, 4, 4, 4, 2,         2, 5, 5, 5, 5, 5, 2, 2, 2, 2, 2, 2, 1), a_reasig = c(0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0), round = structure(c(1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L), class = "factor", .Label = "4")), row.names = c(1L,     52L, 86L, 87L, 112L, 113L, 150L, 151L, 152L, 153L, 178L,     179L, 205L, 206L, 207L, 208L, 209L, 234L, 235L, 279L, 280L,     311L, 312L, 353L), class = "data.frame"), structure(list(        cluster = structure(c(1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         10L, 10L, 10L, 10L, 10L, 10L, 10L, 11L, 11L, 11L, 11L,         11L, 11L, 11L, 11L, 11L, 11L, 11L, 12L, 12L, 12L, 12L,         12L, 12L, 12L, 12L, 13L, 13L, 13L, 13L, 13L, 13L, 13L,         13L, 13L, 14L, 14L, 14L, 14L, 14L, 14L, 14L, 14L, 14L,         14L, 15L, 15L, 15L, 15L, 15L, 15L, 15L, 15L, 15L, 16L,         16L, 16L, 16L, 16L, 16L, 16L, 16L, 17L, 17L, 17L, 17L,         17L, 17L, 17L, 17L, 17L, 18L, 18L, 18L, 18L, 18L, 18L,         18L, 18L, 18L, 18L, 19L, 19L, 19L, 19L, 19L, 2L, 2L,         2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 20L, 20L, 20L,         20L, 20L, 3L, 3L, 3L, 3L, 3L, 3L, 4L, 4L, 4L, 4L, 4L,         4L, 4L, 4L, 4L, 4L, 4L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 6L,         6L, 7L, 7L, 7L, 7L, 7L, 7L, 7L, 7L, 7L, 7L, 8L, 8L, 8L,         8L, 8L, 8L, 8L, 8L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L        ), .Label = c("1", "2", "3", "4", "5", "6", "7", "8",         "9", "10", "11", "12", "13", "14", "15", "16", "17",         "18", "19", "20"), class = "factor"), x = c(574708L,         574110L, 575262L, 574211L, 572927L, 572865L, 572877L,         576075L, 566818L, 566235L, 567529L, 565770L, 569048L,         565652L, 569298L, 579510L, 579319L, 579295L, 578453L,         578632L, 579248L, 578027L, 578997L, 580554L, 577803L,         579845L, 572992L, 574079L, 573550L, 572892L, 574668L,         573053L, 573927L, 574459L, 564243L, 564877L, 564933L,         564761L, 565147L, 564595L, 564194L, 565375L, 563899L,         574717L, 574081L, 575493L, 575999L, 575078L, 574366L,         575930L, 573054L, 573187L, 573425L, 562480L, 560849L,         560799L, 560936L, 562453L, 560438L, 562157L, 560022L,         560314L, 562113L, 562121L, 561633L, 561662L, 560889L,         562384L, 560584L, 560909L, 568063L, 568475L, 568566L,         567142L, 568092L, 566134L, 568897L, 568965L, 565499L,         572441L, 572703L, 572232L, 572151L, 573409L, 572415L,         572420L, 573979L, 574138L, 570607L, 567363L, 566139L,         567148L, 566225L, 567419L, 570631L, 570734L, 569537L,         570771L, 570505L, 569759L, 570560L, 571068L, 570934L,         567783L, 571336L, 570399L, 583456L, 583027L, 584728L,         582374L, 582236L, 576833L, 576000L, 576087L, 575736L,         576783L, 577687L, 571705L, 571777L, 572444L, 572378L,         571352L, 572824L, 572314L, 570568L, 570049L, 572095L,         573238L, 561941L, 560549L, 560458L, 561309L, 560152L,         561009L, 561500L, 583820L, 583906L, 563664L, 563416L,         564229L, 564247L, 564110L, 562798L, 564982L, 563125L,         563733L, 561721L, 579115L, 579932L, 580440L, 580653L,         579892L, 580783L, 577836L, 580086L, 584156L, 584600L,         584078L, 584358L, 583747L, 583042L, 584777L, 584653L,         584142L), y = c(6158695L, 6157862L, 6158567L, 6157318L,         6158668L, 6159256L, 6157323L, 6157916L, 6136853L, 6137700L,         6138532L, 6137019L, 6136344L, 6136509L, 6136245L, 6158101L,         6157180L, 6156886L, 6158627L, 6157029L, 6159149L, 6157956L,         6159193L, 6158751L, 6158997L, 6156219L, 6154323L, 6154126L,         6153627L, 6153844L, 6154127L, 6155457L, 6153024L, 6153089L,         6153696L, 6153371L, 6152814L, 6152525L, 6152624L, 6154968L,         6151795L, 6154876L, 6155377L, 6148486L, 6148033L, 6147332L,         6147740L, 6149822L, 6149889L, 6147054L, 6147964L, 6149392L,         6149790L, 6148527L, 6148748L, 6148845L, 6148019L, 6149985L,         6148854L, 6150143L, 6148375L, 6150021L, 6159382L, 6159358L,         6159849L, 6158575L, 6158721L, 6159971L, 6159228L, 6157801L,         6146680L, 6146299L, 6147037L, 6145445L, 6145448L, 6147644L,         6147277L, 6147096L, 6146162L, 6144230L, 6144693L, 6144312L,         6144900L, 6143920L, 6143403L, 6143131L, 6144066L, 6143315L,         6144532L, 6156524L, 6157403L, 6159077L, 6158525L, 6159425L,         6151327L, 6151192L, 6152460L, 6151843L, 6150154L, 6149756L,         6149845L, 6150274L, 6150069L, 6151563L, 6152135L, 6149501L,         6155159L, 6154210L, 6155733L, 6156153L, 6154310L, 6137069L,         6137431L, 6136937L, 6137257L, 6136303L, 6139046L, 6140899L,         6139970L, 6141120L, 6141379L, 6139497L, 6140663L, 6139631L,         6142415L, 6142020L, 6139320L, 6141063L, 6139692L, 6140044L,         6138425L, 6137794L, 6140139L, 6137573L, 6137391L, 6137022L,         6136833L, 6143911L, 6144032L, 6142408L, 6143975L, 6144384L,         6142125L, 6142698L, 6144776L, 6141302L, 6143151L, 6145175L,         6144940L, 6146168L, 6146382L, 6147273L, 6144628L, 6146210L,         6143957L, 6148872L, 6149102L, 6148181L, 6150071L, 6147831L,         6148192L, 6147920L, 6150666L, 6150811L), weight = c(0.513871377101168,         0.550901632523164, 0.344008838292211, 0.165151672204956,         0.785474966745824, 0.122817386407405, 0.536066253902391,         0.0926152556203306, 0.924321972532198, 0.697539139771834,         0.691467969212681, 0.00692630559206009, 0.823847733670846,         0.645295222522691, 0.894769543781877, 0.995220626704395,         0.489613386103883, 0.756480398355052, 0.402426635846496,         0.491495060734451, 0.808534399606287, 0.00841581495478749,         0.478797567542642, 0.192449602531269, 0.474458581302315,         0.6747138237115, 0.184111420530826, 0.016930561279878,         0.639116713311523, 0.661907109664753, 0.490576893556863,         0.607652891427279, 0.833041093312204, 0.0572722917422652,         0.808186251204461, 0.674185030627996, 0.43124078749679,         0.568660004995763, 0.979439042042941, 0.974753028713167,         0.93427948304452, 0.244842550950125, 0.00402510561980307,         0.710019503952935, 0.50130332633853, 0.522521095816046,         0.683349017286673, 0.962643933249637, 0.803876974852756,         0.102958375355229, 0.695740473922342, 0.263034665025771,         0.17192055657506, 0.308000636519864, 0.835933107417077,         0.457725290907547, 0.435304759768769, 0.0276657671201974,         0.495963071240112, 0.0499963860493153, 0.34112263773568,         0.876832793699577, 0.372668533120304, 0.0616557442117482,         0.341778698377311, 0.0779162363614887, 0.423220758792013,         0.0101891059894115, 0.54344973131083, 0.05437918798998,         0.305039795348421, 0.741616033948958, 0.270497706951573,         0.0494841684121639, 0.0988156476523727, 0.496670980239287,         0.13574977288954, 0.278881686739624, 0.936266565695405,         0.804893916472793, 0.994801041902974, 0.826343502383679,         0.282878895755857, 0.354557034792379, 0.886836963705719,         0.065801584161818, 0.0700161112472415, 0.458008688176051,         0.163674774114043, 0.497440529288724, 0.844189118826762,         0.111577310366556, 0.711540051503107, 0.714574838988483,         0.442606378346682, 0.829872409580275, 0.201432173606008,         0.954502505715936, 0.101857353700325, 0.165606967173517,         0.403435789979994, 0.870417287340388, 0.760741410544142,         0.341528809629381, 0.303296324564144, 0.185100191272795,         0.876653333427384, 0.851631827419624, 0.173997121863067,         0.14441987988539, 0.0180320183280855, 0.536284217378125,         0.221669991733506, 0.943066326668486, 0.229939175304025,         0.963539153337479, 0.322630435461178, 0.488967329496518,         0.109345134114847, 0.318667600397021, 0.877297771861777,         0.958990694722161, 0.712349547771737, 0.59816611232236,         0.934155006660149, 0.702512428862974, 0.181869816035032,         0.207266557728872, 0.851687024347484, 0.987039552303031,         0.522267029387876, 0.883681680075824, 0.510550401872024,         0.313401057617739, 0.825591333443299, 0.721628178376704,         0.444316111737862, 0.467472768621519, 0.575332742882892,         0.460585955297574, 0.539787559770048, 0.317555702291429,         0.313129110494629, 0.114318908425048, 0.422433662228286,         0.511541903950274, 0.528445455478504, 0.994265923742205,         0.452804629458115, 0.62298089871183, 0.413058464415371,         0.985724040539935, 0.153912809677422, 0.270120591158047,         0.685387815814465, 0.440502615878358, 0.0318583238404244,         0.747600480215624, 0.957361328881234, 0.657257393933833,         0.154836883535609, 0.58563402784057, 0.157992691267282,         0.664872472174466), medianx = c(574211, 574211, 574211,         574211, 574211, 574211, 574211, 574211, 567445, 567445,         567445, 567445, 567445, 567445, 567445, 579295, 579295,         579295, 579295, 579295, 579295, 579295, 579295, 579295,         579295, 579295, 573550, 573550, 573550, 573550, 573550,         573550, 573550, 573550, 564218.5, 564218.5, 564218.5,         564218.5, 564218.5, 564218.5, 564218.5, 564218.5, 564218.5,         574897.5, 574897.5, 574897.5, 574897.5, 574897.5, 574897.5,         574897.5, 574897.5, 574897.5, 574897.5, 561820, 561820,         561820, 561820, 561820, 561820, 561820, 561820, 561820,         561662, 561662, 561662, 561662, 561662, 561662, 561662,         561662, 567379, 567379, 567379, 567379, 567379, 567379,         567379, 567379, 567379, 572572, 572572, 572572, 572572,         572572, 572572, 572572, 572572, 572572, 572572, 567419,         567419, 567419, 567419, 567419, 569648, 569648, 569648,         569648, 569648, 569648, 569648, 569648, 569648, 569648,         569648, 569648, 583456, 583456, 583456, 583456, 583456,         576698, 576698, 576698, 576698, 576698, 576698, 571352,         571352, 571352, 571352, 571352, 571352, 571352, 571352,         571352, 571352, 571352, 561500, 561500, 561500, 561500,         561500, 561500, 561500, 583850, 583850, 563664, 563664,         563664, 563664, 563664, 563664, 563664, 563664, 563664,         563664, 579474, 579474, 579474, 579474, 579474, 579474,         579474, 579474, 583991.5, 583991.5, 583991.5, 583991.5,         583991.5, 583991.5, 583991.5, 583991.5, 583991.5), mediany = c(6158546,         6158546, 6158546, 6158546, 6158546, 6158546, 6158546,         6158546, 6136936, 6136936, 6136936, 6136936, 6136936,         6136936, 6136936, 6157956, 6157956, 6157956, 6157956,         6157956, 6157956, 6157956, 6157956, 6157956, 6157956,         6157956, 6154323, 6154323, 6154323, 6154323, 6154323,         6154323, 6154323, 6154323, 6153533.5, 6153533.5, 6153533.5,         6153533.5, 6153533.5, 6153533.5, 6153533.5, 6153533.5,         6153533.5, 6148460, 6148460, 6148460, 6148460, 6148460,         6148460, 6148460, 6148460, 6148460, 6148460, 6148796.5,         6148796.5, 6148796.5, 6148796.5, 6148796.5, 6148796.5,         6148796.5, 6148796.5, 6148796.5, 6159228, 6159228, 6159228,         6159228, 6159228, 6159228, 6159228, 6159228, 6146726,         6146726, 6146726, 6146726, 6146726, 6146726, 6146726,         6146726, 6146726, 6144422, 6144422, 6144422, 6144422,         6144422, 6144422, 6144422, 6144422, 6144422, 6144422,         6157651, 6157651, 6157651, 6157651, 6157651, 6151259.5,         6151259.5, 6151259.5, 6151259.5, 6151259.5, 6151259.5,         6151259.5, 6151259.5, 6151259.5, 6151259.5, 6151259.5,         6151259.5, 6155159, 6155159, 6155159, 6155159, 6155159,         6137510, 6137510, 6137510, 6137510, 6137510, 6137510,         6140899, 6140899, 6140899, 6140899, 6140899, 6140899,         6140899, 6140899, 6140899, 6140899, 6140899, 6139314,         6139314, 6139314, 6139314, 6139314, 6139314, 6139314,         6138685, 6138685, 6143151, 6143151, 6143151, 6143151,         6143151, 6143151, 6143151, 6143151, 6143151, 6143151,         6145718, 6145718, 6145718, 6145718, 6145718, 6145718,         6145718, 6145718, 6148987, 6148987, 6148987, 6148987,         6148987, 6148987, 6148987, 6148987, 6148987), dist = c(518.854507545227,         691.416661644771, 1051.20977925436, 1228, 1289.78292747268,         1521.78053608265, 1809.77484787473, 1967.5863386393,         632.46976212306, 1431.01222915809, 1598.208997597, 1677.05515711321,         1708.82210893937, 1843.14351041909, 1977.6475924694,         259.326049597799, 776.371045312742, 1070, 1076.66382868563,         1139.69206367334, 1193.92545830969, 1268, 1272.38869847229,         1488.99496305394, 1819.27045817822, 1821.9958836397,         558, 564.490921096168, 696, 813.882669676655, 1135.05065966238,         1238.1296378005, 1352.6011976928, 1532.65684352369, 164.33654493143,         678.254008465855, 1013.99728796481, 1145.1543564079,         1299.73170308337, 1483.08546618191, 1738.67262588447,         1771.94765724047, 1870.98169419158, 182.362962248369,         921.412638289708, 1275.54076767464, 1315.94158304995,         1373.90838486414, 1524.64200716103, 1744.38878980576,         1909.05951976359, 1947.93076108983, 1984.22686454952,         712.902693219769, 972.210496754689, 1022.15128528022,         1177.26897946051, 1346.55904066625, 1383.19566584052,         1388.03142975943, 1846.74477121231, 1940.98847240266,         476.567938493558, 477.054504223574, 621.676764886705,         653, 924.433880815713, 1036.01785698896, 1078, 1613.48628751533,         685.545038637142, 1176.24189689026, 1227.06560541806,         1302.73942137328, 1463.43875854099, 1546.85131800054,         1614.90711807212, 1628.58711771891, 1962.77762367518,         232.432785983389, 301.001661124984, 357.351367704113,         636.965462172008, 975.998463113544, 1031.02376306271,         1299.91730506213, 1451.33903688973, 1917.76041256461,         1968.07647209147, 1128.39044660968, 1303.80366620132,         1451.5223043412, 1479.69996958843, 1774, 985.314797412482,         1088.09569891623, 1205.62069076472, 1265.54385542343,         1398.77777005499, 1507.59187116408, 1683.0193849151,         1728.47049439671, 1752.45149718901, 1889.53360647542,         1901.53733857634, 1912.15147151056, 0, 1041.46147312323,         1395.51424213442, 1469.27192854148, 1486.33811765695,         461.200607111482, 702.456404341223, 837.645509747411,         994.712521284416, 1209.98925615065, 1826.85987421039,         353, 1021.59972592009, 1114.13868077542, 1132.72944695545,         1402, 1490.79844378776, 1591.62432753461, 1706.72551981858,         1718.85136064757, 1745.07592958014, 1893.11700642089,         580.831300809452, 1198.87488921905, 1369.70252244785,         1531.95332827081, 1580.42051366084, 1808.91182759138,         1923, 1663.27057329828, 1852.8464588303, 760, 915.24040557659,         933.420591159205, 1009.38842870324, 1311.18457892091,         1342.62131667868, 1393.67607427264, 1712.05899431065,         1850.2870047644, 1943, 650.945466225858, 902.800088613199,         1065.67161921485, 1353.12120669214, 1610.20154018061,         1703.40277092648, 1710.29471144595, 1864.31354659027,         200.711858144954, 619.271547868946, 810.628305698734,         1144.2806692416, 1181.57363291502, 1238.37605354755,         1324.95254632006, 1804.61166182644, 1830.19841820498),         orden_dist = c(1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 1L, 2L,         3L, 4L, 5L, 6L, 7L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L,         10L, 11L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 1L, 2L, 3L,         4L, 5L, 6L, 7L, 8L, 9L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L,         9L, 10L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 1L, 2L,         3L, 4L, 5L, 6L, 7L, 8L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L,         9L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L, 1L, 2L,         3L, 4L, 5L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L,         11L, 12L, 1L, 2L, 3L, 4L, 5L, 1L, 2L, 3L, 4L, 5L, 6L,         1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L, 11L, 1L, 2L,         3L, 4L, 5L, 6L, 7L, 1L, 2L, 1L, 2L, 3L, 4L, 5L, 6L, 7L,         8L, 9L, 10L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 1L, 2L,         3L, 4L, 5L, 6L, 7L, 8L, 9L), max_n_cl = c(8L, 8L, 8L,         8L, 8L, 8L, 8L, 8L, 7L, 7L, 7L, 7L, 7L, 7L, 7L, 11L,         11L, 11L, 11L, 11L, 11L, 11L, 11L, 11L, 11L, 11L, 8L,         8L, 8L, 8L, 8L, 8L, 8L, 8L, 9L, 9L, 9L, 9L, 9L, 9L, 9L,         9L, 9L, 10L, 10L, 10L, 10L, 10L, 10L, 10L, 10L, 10L,         10L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 8L, 8L, 8L,         8L, 8L, 8L, 8L, 8L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L,         10L, 10L, 10L, 10L, 10L, 10L, 10L, 10L, 10L, 10L, 5L,         5L, 5L, 5L, 5L, 12L, 12L, 12L, 12L, 12L, 12L, 12L, 12L,         12L, 12L, 12L, 12L, 5L, 5L, 5L, 5L, 5L, 6L, 6L, 6L, 6L,         6L, 6L, 11L, 11L, 11L, 11L, 11L, 11L, 11L, 11L, 11L,         11L, 11L, 7L, 7L, 7L, 7L, 7L, 7L, 7L, 2L, 2L, 10L, 10L,         10L, 10L, 10L, 10L, 10L, 10L, 10L, 10L, 8L, 8L, 8L, 8L,         8L, 8L, 8L, 8L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L, 9L),         a_reasig = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,         0, 0, 0, 0, 0, 0), round = structure(c(1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,         1L, 1L, 1L, 1L, 1L, 1L), class = "factor", .Label = "5")), row.names = c(1L,     2L, 3L, 4L, 5L, 6L, 7L, 8L, 14L, 15L, 16L, 17L, 18L, 19L,     20L, 28L, 29L, 30L, 31L, 32L, 33L, 34L, 35L, 36L, 37L, 38L,     53L, 54L, 55L, 56L, 57L, 58L, 59L, 60L, 72L, 73L, 74L, 75L,     76L, 77L, 78L, 79L, 80L, 90L, 91L, 92L, 93L, 94L, 95L, 96L,     97L, 98L, 99L, 112L, 113L, 114L, 115L, 116L, 117L, 118L,     119L, 120L, 128L, 129L, 130L, 131L, 132L, 133L, 134L, 135L,     139L, 140L, 141L, 142L, 143L, 144L, 145L, 146L, 147L, 158L,     159L, 160L, 161L, 162L, 163L, 164L, 165L, 166L, 167L, 184L,     185L, 186L, 187L, 188L, 201L, 202L, 203L, 204L, 205L, 206L,     207L, 208L, 209L, 210L, 211L, 212L, 225L, 226L, 227L, 228L,     229L, 238L, 239L, 240L, 241L, 242L, 243L, 261L, 262L, 263L,     264L, 265L, 266L, 267L, 268L, 269L, 270L, 271L, 280L, 281L,     282L, 283L, 284L, 285L, 286L, 295L, 296L, 312L, 313L, 314L,     315L, 316L, 317L, 318L, 319L, 320L, 321L, 327L, 328L, 329L,     330L, 331L, 332L, 333L, 334L, 352L, 353L, 354L, 355L, 356L,     357L, 358L, 359L, 360L), class = "data.frame"), check.names = TRUE,     fix.empty.names = TRUE, stringsAsFactors = TRUE): arguments imply differing number of rows: 46, 32, 27, 24, 164