Calulate EPP score for all models

Calulate EPP score for all models

calculate_epp(results, decreasing_metric = TRUE,
  compare_in_split = TRUE, keep_data = FALSE)

Arguments

results
data frame with results for one dataset. Data should be in the following format. First 3 columns should correspond to: model, split, score. See more in 'details' section.

decreasing_metric
Logical. If TRUE used metric is considered as decreasing, that means a model with higher score value is considered as better model. If FALSE used metric will be considered as increasing.

compare_in_split
Logical. If TRUE compares models only in the same fold. If FALSE compares models across folds.

keep_data
Logical. If FALSE only EPP scores will be returned. If TRUE original data frame with new 'epp' column will be returned.

Details

Format of the data frame passed via results parameter. First column should correspond to a model. Dofferent settings of hyperparameters of the same model should have different values in this column. Second column corresponds to indexes of splits. As EPP is based on Elo rating system, power of model is assessed by comparing its results with other models on multiple data splits. Therefore, each model should be evaluated on multiple train-test splits. Indexes of splits should be in this column. And should match across models. Third column contains score used to evaluate models. It can be both decreasing or increasing metric, just remember to set the decreasing_metric parameter accordingly. The following columns can be of any kind.

Examples

library(epp)
data(auc_scores)
calculate_epp(auc_scores)

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#>              model          epp
#> 1       catboost_1  -0.80310655
#> 2       catboost_2   2.92737911
#> 3       catboost_3  -2.06876823
#> 4            gbm_1 -34.17937952
#> 5           gbm_10   1.67375851
#> 6           gbm_11   2.72549809
#> 7            gbm_2   1.26754392
#> 8            gbm_3  -0.17923285
#> 9            gbm_4 -32.60978901
#> 10           gbm_5 -21.46324830
#> 11           gbm_6 -28.57294825
#> 12           gbm_7   1.11048994
#> 13           gbm_8 -19.61875543
#> 14           gbm_9   1.29448886
#> 15        glmnet_1 -19.89176506
#> 16       glmnet_10 -22.19062878
#> 17       glmnet_11 -44.36998857
#> 18        glmnet_2 -44.36998857
#> 19        glmnet_3 -44.36998857
#> 20        glmnet_4 -25.67254948
#> 21        glmnet_5 -44.36998857
#> 22        glmnet_6 -27.96898698
#> 23        glmnet_7 -30.45058970
#> 24        glmnet_8 -44.36998857
#> 25        glmnet_9 -44.36998857
#> 26          kknn_1 -25.24587475
#> 27         kknn_10 -25.24587475
#> 28         kknn_11 -23.39728197
#> 29          kknn_2 -21.98948630
#> 30          kknn_3 -27.90557132
#> 31          kknn_4 -23.70831004
#> 32          kknn_5 -22.06787186
#> 33          kknn_6 -22.21708876
#> 34          kknn_7 -22.27907377
#> 35          kknn_8 -30.21374205
#> 36          kknn_9 -30.21374205
#> 37  randomForest_1   2.14555397
#> 38 randomForest_10   0.32058577
#> 39 randomForest_11   2.02105346
#> 40  randomForest_2   1.55332801
#> 41  randomForest_3   1.09446546
#> 42  randomForest_4   2.25975140
#> 43  randomForest_5   2.11707813
#> 44  randomForest_6   1.98548542
#> 45  randomForest_7   2.22399978
#> 46  randomForest_8   1.89287613
#> 47  randomForest_9   1.47569672
#> 48        ranger_1   1.41592697
#> 49       ranger_10   2.64257254
#> 50       ranger_11   3.07750846
#> 51        ranger_2   2.18474674
#> 52        ranger_3   2.71790594
#> 53        ranger_4   2.91546845
#> 54        ranger_5   3.59917250
#> 55        ranger_6   2.71032495
#> 56        ranger_7   3.26102899
#> 57        ranger_8   3.40568901
#> 58        ranger_9   3.71846154
#> 59           svm_1 -29.06884202
#> 60          svm_10 -29.04998649
#> 61          svm_11 -30.31969406
#> 62           svm_2 -35.61717026
#> 63           svm_3 -38.90144358
#> 64           svm_4 -29.57902149
#> 65           svm_5 -40.46716846
#> 66           svm_6 -37.16719006
#> 67           svm_7 -26.95669190
#> 68           svm_8 -27.96898698
#> 69           svm_9 -22.64182875
#> 70       xgboost_1 -29.94863311
#> 71      xgboost_10 -20.02570240
#> 72      xgboost_11   3.13185143
#> 73       xgboost_2 -19.99519574
#> 74       xgboost_3   0.24708631
#> 75       xgboost_4 -20.00538993
#> 76       xgboost_5   1.14632009
#> 77       xgboost_6   1.81065306
#> 78       xgboost_7 -19.91267808
#> 79       xgboost_8 -20.86121390
#> 80       xgboost_9  -0.07265424