Function for iterative clustering

The function clusters population to create service's centers. Iterates EPP::clust_pop for the remaining population in each step. It allows to define two distances of service and two group sizes for several rounds of iteration.

clust_it(pop, m = 5, l = 4, g1 = 5, g2 = g1 * 0.5, d1 = 1000, d2 = d1 * 2)

Arguments

pop

Population to attend (dataframe with three variables: x, y, and weight). x and y are plain coordinates in the defined CRS

m

Number of iteration rounds. Default 5

l

Number of iteration rounds with the first group size (g1). Default 4

g1

Size of the groups for the first l iterations. Default 50

g2

Size of the groups for the last m-l iterations. Default g1 * 0.5

d1

Distance range of service for the first iterations Default 1000

d2

Second distance range of service. Default double of d1

Value

Return a LIST with:

Clustered

Population assigned to created centers by clusterization

pop

Remaining non-assigned population

Examples

clu <- clust_it(pop = pop_epp)